Separate PlayerViewControllerRepresentable
